requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.0
pydantic[email]>=2.0.0
python-dotenv>=1.0.0
aiofiles>=23.0.0
pymysql>=1.1.0
cryptography>=42.0.0
passlib[bcrypt]>=1.7.4
bcrypt<4.0
python-jose[cryptography]>=3.3.0
authlib>=1.3.0
httpx>=0.27.0
itsdangerous>=2.1.0
slowapi>=0.1.9
